requirements.txt
transformers
Pillow
requests
py-cord
torch
torchvision
torchaudio
python-dotenv
websocket-client
